[vulkan] Refactored param descriptors
