+++ drape/mesh_object.hpp
@@ -106,7 +106,7 @@ private:
   };
 
   void InitForOpenGL();
-  void InitForVulkan();
+  void InitForVulkan(ref_ptr<dp::GraphicsContext> context);
 
 #if defined(OMIM_METAL_AVAILABLE)
   // Definition of this method is in a .mm-file.
