+++ drape/vulkan/vulkan_mesh_object_impl.cpp
@@ -2,6 +2,7 @@
 #include "drape/pointers.hpp"
 #include "drape/vulkan/vulkan_base_context.hpp"
 #include "drape/vulkan/vulkan_staging_buffer.hpp"
+#include "drape/vulkan/vulkan_param_descriptor.hpp"
 #include "drape/vulkan/vulkan_utils.hpp"
 
 #include "base/assert.hpp"
@@ -33,16 +34,15 @@ VkPrimitiveTopology GetPrimitiveType(MeshObject::DrawPrimitive primitive)
 class VulkanMeshObjectImpl : public MeshObjectImpl
 {
 public:
-  explicit VulkanMeshObjectImpl(ref_ptr<dp::MeshObject> mesh)
+  VulkanMeshObjectImpl(ref_ptr<VulkanObjectManager> objectManager, ref_ptr<dp::MeshObject> mesh)
     : m_mesh(std::move(mesh))
+    , m_objectManager(objectManager)
+    , m_descriptorUpdater(objectManager)
   {}
 
   void Build(ref_ptr<dp::GraphicsContext> context, ref_ptr<dp::GpuProgram> program) override
   {
-    ref_ptr<dp::vulkan::VulkanBaseContext> vulkanContext = context;
-    m_objectManager = vulkanContext->GetObjectManager();
-
-    ResetDescriptorSetGroup();
+    m_descriptorUpdater.Reset();
     m_geometryBuffers.resize(m_mesh->m_buffers.size());
     m_bindingInfoCount = static_cast<uint8_t>(m_mesh->m_buffers.size());
     CHECK_LESS_OR_EQUAL(m_bindingInfoCount, kMaxBindingInfo, ());
@@ -74,7 +74,7 @@ public:
 
   void Reset() override
   {
-    ResetDescriptorSetGroup();
+    m_descriptorUpdater.Reset();
     for (auto const & b : m_geometryBuffers)
       m_objectManager->DestroyObject(b);
     m_geometryBuffers.clear();
@@ -151,13 +151,13 @@ public:
     vulkanContext->SetPrimitiveTopology(GetPrimitiveType(m_mesh->m_drawPrimitive));
     vulkanContext->SetBindingInfo(m_bindingInfo, m_bindingInfoCount);
 
-    UpdateDescriptorSetGroup(vulkanContext);
+    m_descriptorUpdater.Update(context);
+    auto descriptorSet = m_descriptorUpdater.GetDescriptorSet();
 
     uint32_t dynamicOffset = vulkanContext->GetCurrentDynamicBufferOffset();
     vkCmdBindDescriptorSets(commandBuffer, VK_PIPELINE_BIND_POINT_GRAPHICS,
                             vulkanContext->GetCurrentPipelineLayout(), 0, 1,
-                            &m_descriptorSetGroups[m_descriptorSetIndex].m_descriptorSet,
-                            1, &dynamicOffset);
+                            &descriptorSet, 1, &dynamicOffset);
 
     vkCmdBindPipeline(commandBuffer, VK_PIPELINE_BIND_POINT_GRAPHICS,
                       vulkanContext->GetCurrentPipeline());
@@ -173,53 +173,19 @@ public:
   void Unbind() override {}
 
 private:
-  void ResetDescriptorSetGroup()
-  {
-    for (auto const & g : m_descriptorSetGroups)
-      m_objectManager->DestroyDescriptorSetGroup(g);
-    m_descriptorSetGroups.clear();
-  }
-
-  void UpdateDescriptorSetGroup(ref_ptr<dp::vulkan::VulkanBaseContext> vulkanContext)
-  {
-    if (m_program != vulkanContext->GetCurrentProgram())
-    {
-      ResetDescriptorSetGroup();
-      m_program = vulkanContext->GetCurrentProgram();
-    }
-
-    if (m_updateDescriptorFrame != vulkanContext->GetCurrentFrameIndex())
-    {
-      m_updateDescriptorFrame = vulkanContext->GetCurrentFrameIndex();
-      m_descriptorSetIndex = 0;
-    }
-    else
-    {
-      m_descriptorSetIndex++;
-    }
-
-    CHECK_LESS_OR_EQUAL(m_descriptorSetIndex, m_descriptorSetGroups.size(), ());
-    if (m_descriptorSetIndex == m_descriptorSetGroups.size())
-      m_descriptorSetGroups.emplace_back(m_objectManager->CreateDescriptorSetGroup(m_program));
-
-    m_descriptorSetGroups[m_descriptorSetIndex].Update(vulkanContext->GetDevice(),
-                                                       vulkanContext->GetCurrentParamDescriptors());
-  }
-
   ref_ptr<dp::MeshObject> m_mesh;
   ref_ptr<VulkanObjectManager> m_objectManager;
   std::vector<VulkanObject> m_geometryBuffers;
   BindingInfoArray m_bindingInfo;
   uint8_t m_bindingInfoCount = 0;
-  std::vector<DescriptorSetGroup> m_descriptorSetGroups;
-  ref_ptr<VulkanGpuProgram> m_program;
-  uint32_t m_updateDescriptorFrame = 0;
-  uint32_t m_descriptorSetIndex = 0;
+  ParamDescriptorUpdater m_descriptorUpdater;
 };
 }  // namespace vulkan
 
-void MeshObject::InitForVulkan()
+void MeshObject::InitForVulkan(ref_ptr<dp::GraphicsContext> context)
 {
-  m_impl = make_unique_dp<vulkan::VulkanMeshObjectImpl>(make_ref(this));
+  ref_ptr<dp::vulkan::VulkanBaseContext> vulkanContext = context;
+  m_impl = make_unique_dp<vulkan::VulkanMeshObjectImpl>(vulkanContext->GetObjectManager(),
+                                                        make_ref(this));
 }
 }  // namespace dp
